scripts/requirements.txt
# Fragua Systems — Script Dependencies
#
# Install: pip install -r scripts/requirements.txt

Pillow>=10.0.0
